oleksinski_siakala/requirements.txt
keras==2.15.0
keras_contrib==2.0.8
matplotlib==3.8.2
numpy==1.26.4
opencv_python==4.9.0.80
Pillow==10.2.0
tqdm==4.66.1
